backend/requirements.txt
fastapi
pyjwt
sqlalchemy
pg8000
uvicorn
python-dotenv
pydantic
passlib
